Add support for multi-config generators
An example for such a generator is `Ninja Multi-Config` (introduced in
CMake 3.17).

The commit changes the semantic (and name) of `target-name`
variables/functions to `target` which is now a combination of the
actual _target name_ and the _configuration name_ separated by
`cmake-integration--multi-config-separator` (in the - now - special
case of a _non multi-config_ generator, the _configuration name_ is
nil and _target_ is identical to _target name_).

The handling of the `all` target has changed, too. It is now included
in the list you get from
`cmake-integration-get-cmake-targets-from-codemodel-json-file` (since
it also needs to have the _multi-config_ variants).

In addition to the above, I also added a call to
`check-if-build-folder-exists-and-throws-if-not` in
`cmake-integration-run-last-target`.

Closes #3.