Allow specifying the working directory when running a target
The `cmake-integration-run-working-directory` variable was introduced
to allow specifying the working directory when executing a target with
the `cmake-integration-run-last-target` command.

The possible values for `cmake-integration-run-working-directory` are
'root, 'build, 'bin, or "some string". The first three values are
symbols denoting the project root folder, the build folder and the
folder containing the executable, respectively. The fourth value can
be any string and it is interpreted as a subfolder of the project root
folder (this subfolder must already exist).