Cache target type
If your project has a lot of targets, calling
`cmake-integration-save-and-compile` can be slow just because we need
to open and parse a single json file for each target in order to get
the target type. This commit add a cache of the target type in a hash
table such that the second time you call
`cmake-integration-save-and-compile` this process is avoided.