Fix cmake-integration-get-build-folder when parent preset is hidden
When presets are used and the preset does not have the `binaryDir`
field, cmake-integration search the parent presets to find the first
one containing a `binaryDir` field. However, if the parent was hidden
it was not searched (by mistake). This is now fixed by adding an
optional `show-hidden` option to `ci-get-configure-presets` and
adapting the rest of the code to make `ci-get-build-folder` work
correctly.