Make ci--get-binaryDir-with-replacements include remote prefix
When working on a remote (with TRAMP), the returned build folder must
include the TRAMP remote prefix (that "/<protocol>:<host>:" part).
This will be the case if the value of

when your preset includes something like the "binaryDir" field starts
with "${sourceDir}" because cmake-integration will do the replacement.
When this is not the case and we are using a remote connection,
`ci--get-binaryDir-with-replacements` will now add the remote prefix
to the returned value.