Allow customizing a launch function
The `cmake-integration-program-launcher` variable was added, which
stores the function that is used to launch the target executable. The
stored function receives two arguments: the command that should be
run, and the name that should be used for a buffer displaying the
running program.

The default value is set to a new
`cmake-integration-default-program-launch-function` function, which
just run the executable in a compile buffer. Hence, this commit does
not change the current behavior, but allow us to customize how the
program is run.