Refactor ci-debug-launcher-function to simplify usage
The `ci-debug-launcher-function` variable has been refactored to
simplify customization. It can now be set to one of two predefined
build symbols representing the available built-in functions, in
addition to supporting custom functions. This change makes it easier
to configure the variable for typical use cases while retaining
flexibility for advanced customization.