Add support for persisting the current state
Add `cmake-integration-persistence` with functions to save (cmake-integration-save-state) and restore (cmake-integration-restore-state) the current state in cmake-integration.

The function `cmake-integration-should-restore-state-p` can be used to indicate if there is stated saved that should be restored. This can be used in hooks or even in a lambda function set to a keybinding to determine if the current state should be restored with `cmake-integration-restore-state`. The function `cmake-integration-maybe-restore-state` does exactly that.