Always create link to `compile_commands.json` if flag is set
If `ci-create-compile-commands-link` is set to `t`, ensure that a
symbolic link to `compile_commands.json` is created, even if the file
does not yet exist.

Rationale:

When `ci--create-link-to-compile-commands` is invoked during the
initial compilation, the build process, including the creation of the
build folder, is still ongoing. At this stage, CMake may not have
generated the `compile_commands.json` file yet. If the link creation
is conditional on the file's existence, the symbolic link will not be
created until a subsequent compilation event after the build
completes.

By forcing the creation of the link regardless of whether
`compile_commands.json` currently exists, we ensure that the link is
available as soon as the file is generated by CMake, avoiding any
delay in subsequent compilation processes that depend on
`compile_commands.json` being linked correctly.